Extract string_2_perspective_snapshot()
