+++ cli/src/main.rs
@@ -27,6 +27,8 @@ use serde_json::Value;
 use startup::executor_data_path;
 use util::{maybe_parse_datetime, readline_masked};
 
+use crate::util::string_2_perspective_snapshot;
+
 /// AD4M command line interface.
 /// Provides all means of interacting with the AD4M executor / agent.
 /// See help of commands for more information.
@@ -546,32 +548,8 @@ async fn main() -> Result<()> {
                     println!("{:?}", result);
                 },
                 RuntimeFunctions::SetStatus { status } => {
-                    use rand::{thread_rng, Rng};
-                    use rand::distributions::Alphanumeric;
-
-
-                    let rand_name: String = thread_rng()
-                        .sample_iter(&Alphanumeric)
-                        .take(30)
-                        .map(char::from)
-                        .collect();
-                    let temp_perspective = perspectives::run_add(cap_token.clone(), rand_name).await?;
-                    println!("Created temporary perspective: {}", temp_perspective);
-                    perspectives::run_add_link(
-                        cap_token.clone(), 
-                        temp_perspective.clone(), 
-                        "ad4m://self".to_string(), 
-                        format!("literal://string:{}", urlencoding::encode(&status)), 
-                        None
-                    ).await?;
-                    println!("Added status link to temporary perspective");
-
-                    let snapshot = perspectives::run_snapshot(cap_token.clone(), temp_perspective.clone()).await?;
-                    println!("Created snapshot of temporary perspective");
-
-                    perspectives::run_remove(cap_token.clone(), temp_perspective).await?;
-
-                    runtime::run_set_status(cap_token, snapshot.into()).await?;
+                    let perspective = string_2_perspective_snapshot(cap_token.clone(), status).await?;
+                    runtime::run_set_status(cap_token, perspective.into()).await?;
                     println!("Status set!");
                     
                 },
