+++ cli/src/util.rs
@@ -1,4 +1,4 @@
-use crate::startup::get_executor_url;
+use crate::{startup::get_executor_url, types::Perspective};
 use anyhow::{anyhow, Result};
 use async_tungstenite::tungstenite::{client::IntoClientRequest, http::HeaderValue, Message};
 use futures::StreamExt;
@@ -136,4 +136,34 @@ pub fn readline_masked(prompt: &str) -> Result<String> {
     rl.set_auto_add_history(false); // make sure password is not added to history
     let passwd = rl.readline(prompt)?;
     Ok(passwd)
+}
+
+pub async fn string_2_perspective_snapshot(cap_token: String, string: String) -> Result<Perspective> {
+    use rand::{thread_rng, Rng};
+    use rand::distributions::Alphanumeric;
+    use crate::perspectives;
+
+
+    let rand_name: String = thread_rng()
+        .sample_iter(&Alphanumeric)
+        .take(30)
+        .map(char::from)
+        .collect();
+    let temp_perspective = perspectives::run_add(cap_token.clone(), rand_name).await?;
+    println!("Created temporary perspective: {}", temp_perspective);
+    perspectives::run_add_link(
+        cap_token.clone(), 
+        temp_perspective.clone(), 
+        "ad4m://self".to_string(), 
+        format!("literal://string:{}", urlencoding::encode(&string)), 
+        None
+    ).await?;
+    println!("Added status link to temporary perspective");
+
+    let snapshot = perspectives::run_snapshot(cap_token.clone(), temp_perspective.clone()).await?;
+    println!("Created snapshot of temporary perspective");
+
+    perspectives::run_remove(cap_token.clone(), temp_perspective).await?;
+
+    Ok(snapshot)
 }
